+++ src/main/java/com/prgrms/be/intermark/common/dto/page/dto/PageListIndexSize.java
@@ -1,4 +1,4 @@
-package com.prgrms.be.intermark.common.page.dto;
+package com.prgrms.be.intermark.common.dto.page.dto;
 
 public enum PageListIndexSize {
     ADMIN_PERFORMANCE_LIST_INDEX_SIZE(10);
