+++ src/main/java/com/prgrms/be/intermark/common/dto/page/dto/PageResponseDTO.java
@@ -1,11 +1,12 @@
-package com.prgrms.be.intermark.common.page.dto;
-
-import lombok.Getter;
-import org.springframework.data.domain.Page;
+package com.prgrms.be.intermark.common.dto.page.dto;
 
 import java.util.List;
 import java.util.function.Function;
 import java.util.stream.IntStream;
+
+import org.springframework.data.domain.Page;
+
+import lombok.Getter;
 
 @Getter
 public class PageResponseDTO<E, DTO> {
