+++ src/main/java/com/prgrms/be/intermark/domain/performance/controller/PerformanceApiController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import com.prgrms.be.intermark.common.page.dto.PageResponseDTO;
+import com.prgrms.be.intermark.common.dto.page.dto.PageResponseDTO;
 import com.prgrms.be.intermark.domain.performance.dto.PerformanceCommandResponseDTO;
 import com.prgrms.be.intermark.domain.performance.dto.PerformanceCreateRequestDTO;
 import com.prgrms.be.intermark.domain.performance.dto.PerformanceInfoResponseDTO;
